Fix win32 build (#210)
* Initial attempt at better windows bootstrap

* Lower build requirements to Win7 and VS2013

* Fix Win32 Build

* LLVM for windows may not come with llvm-link.exe. Use the one we build from soruce.
* Enable parallel building on Windows
* Handle Windows' double == long double weirdness
* Fixes for std=g++11 breaking clang-cl

+++ mcsema/Arch/X86/Semantics/Bitcode/FXAM.cpp
@@ -4,7 +4,7 @@
 // FXAM semantics are adaptd from Remill
 DEFINE_SEMANTICS(FXAM)
 {
-  auto st0 = state->ST0;
+  auto st0 = DOUBLE(state->ST0);
 
   uint8_t sign = __builtin_signbit(st0) == 0 ? 0 : 1;
   auto c = __builtin_fpclassify(FP_NAN, FP_INFINITE, FP_NORMAL, FP_SUBNORMAL,
